fix(logging): move dotenv import to fix logging on server-side
